requirements.txt
bcrypt==4.1.3
pymongo==4.7.2
robotframework==7.0
robotframework-browser==18.4.0
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.7
